Refactor: Move Definition related functions from language/utils to middlewareDefinition
