+++ client/src/language/middlewareDefinition.ts
@@ -3,11 +3,11 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { Location, Position, Range, commands, type LocationLink, type TextDocument, workspace } from 'vscode'
+import { Location, Position, Range, commands, type LocationLink, type TextDocument, workspace, type Uri } from 'vscode'
 import { type DefinitionMiddleware } from 'vscode-languageclient'
 
 import { requestsManager } from './RequestManager'
-import { changeDefinitionUri, checkIsDefinitionRangeEqual, checkIsDefinitionUriEqual, convertToSameDefinitionType, getDefinitionUri, getEmbeddedLanguageDocPosition, getOriginalDocRange } from './utils'
+import { getEmbeddedLanguageDocPosition, getOriginalDocRange } from './utils'
 import { logger } from '../lib/src/utils/OutputLogger'
 import { type EmbeddedLanguageDocInfos, embeddedLanguageDocsManager } from './EmbeddedLanguageDocsManager'
 
@@ -46,6 +46,67 @@ export const middlewareProvideDefinition: DefinitionMiddleware['provideDefinitio
     return await processDefinitions(tempResult, document, embeddedLanguageTextDocument, embeddedLanguageDocInfos)
   } else {
     return await processDefinitions(tempResult, document, embeddedLanguageTextDocument, embeddedLanguageDocInfos)
+  }
+}
+
+export const checkIsDefinitionUriEqual = (definition: Location | LocationLink, uri: Uri): boolean => {
+  if (definition instanceof Location) {
+    return definition.uri.fsPath === uri.fsPath
+  }
+  return definition.targetUri.fsPath === uri.fsPath
+}
+
+export const changeDefinitionUri = (definition: Location | LocationLink, uri: Uri): void => {
+  if (definition instanceof Location) {
+    definition.uri = uri
+  } else {
+    definition.targetUri = uri
+  }
+}
+
+export const getDefinitionUri = (definition: Location | LocationLink): Uri => {
+  if (definition instanceof Location) {
+    return definition.uri
+  }
+  return definition.targetUri
+}
+
+export const checkIsDefinitionRangeEqual = (definition: Location | LocationLink, range: Range): boolean => {
+  if (definition instanceof Location) {
+    return definition.range.isEqual(range)
+  }
+  return definition.targetRange.isEqual(range)
+}
+
+export const convertDefinitionToLocation = (definition: Location | LocationLink): Location => {
+  if (definition instanceof Location) {
+    return definition
+  }
+  return {
+    uri: definition.targetUri,
+    range: definition.targetRange
+  }
+}
+
+export const convertDefinitionToLocationLink = (definition: Location | LocationLink): LocationLink => {
+  if (definition instanceof Location) {
+    return {
+      targetUri: definition.uri,
+      targetRange: definition.range,
+      targetSelectionRange: definition.range
+    }
+  }
+  return definition
+}
+
+export const convertToSameDefinitionType = <DefinitionType extends Location | LocationLink>(
+  referenceDefinition: DefinitionType,
+  definitionToConvert: Location | LocationLink
+): DefinitionType => {
+  if (referenceDefinition instanceof Location) {
+    return convertDefinitionToLocation(definitionToConvert) as DefinitionType
+  } else {
+    return convertDefinitionToLocationLink(definitionToConvert) as DefinitionType
   }
 }
